The GUI is now the main file
